Add name banner.
Create the ._banner macro and use it in .2c to draw after the aside is
finished. The macro attempts to handle both graphical and terminal
devices well, but gets extremely messy in the process (and may not
always succeed in achieving this; I've only tested with a couple
different names/titles). I've attempted to include comments for the
particularly confusing aspects, but it probably could use some
refactoring in the future.