Fine-tune spacing.
Increase kerning spacing, the width of the aside, the width of the
banner, and spaces in the section headers and banner name. Increasing
the size of a space is a little awkward because spaces can't be replaced
by .tr and we don't want to change the spacing for _all_ glyphs as with
.tkf. I found the simplest (though not most elegant) solution was to
create the ._inc_space macro which would iterate over a given string and
add \h'' escapes, storing the result in a dedicated _out string. If I
can come up with a nicer solution (ideally one that doesn't create an
extra string), then I may refactor this in the future.