Truncate extra space at the end of resume.txt.
By default (at least on my machine), groff considers terminals to have a
height of 66 lines. My résumé exceeds that, so an extra page is added
which creates a lot of extra space at the bottom of the output. Since
I'm storing this in a text file and don't care about page size, just set
the page length to the length of the output, plus an extra line for
padding.